refactor: 댓글 구조 변경
AS-IS:
PostPage -> PostContent

TO-BE:
PostPage -> PostContent -> CommentBlock
